meta: modularised search results list
