+++ app/page.tsx
@@ -15,6 +15,7 @@ import { toast } from "react-hot-toast";
 
 import SearchBar from "@/components/SearchBar";
 import Button from "@/components/ui/Button";
+import Results from "@/components/Results";
 
 function useDebounce(callback: (t: string) => Promise<void> | void) {
     let timeout: null | NodeJS.Timeout = null;
@@ -30,31 +31,6 @@ function useDebounce(callback: (t: string) => Promise<void> | void) {
         }, 500);
         timeout = t;
     };
-}
-
-function handleLastUpdated(date: Date, currentDate: string) {
-    const now = new Date(currentDate);
-    const diff = now.getTime() - date.getTime();
-    const days = Math.floor(diff / (1000 * 60 * 60 * 24));
-    const hours = Math.floor((diff % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60));
-    const months = Math.floor(days / 30);
-    const years = Math.floor(months / 12);
-
-    if (days === 0) {
-        return `${hours} hours ago`;
-    } else if (days === 1) {
-        return `${days} day ago`;
-    } else if (days < 30) {
-        return `${days} days ago`;
-    } else if (months === 1) {
-        return `${months} month ago`;
-    } else if (months < 12) {
-        return `${months} months ago`;
-    } else if (years === 1) {
-        return `${years} years ago`;
-    } else {
-        return `${years} years ago`;
-    }
 }
 
 export default function Home() {
@@ -190,74 +166,13 @@ export default function Home() {
                         <h2 className="my-10 text-4xl font-semibold ">
                             Search results: {searched}
                         </h2>
-                        <div className="results h-full w-full overflow-y-auto">
-                            {results?.objects.map(result => {
-                                return (
-                                    <div
-                                        key={result.package.name}
-                                        className="result border-b-2 py-5"
-                                    >
-                                        <h2 className="text-2xl font-semibold tracking-wider">
-                                            {result.package.name}
-                                            <a
-                                                href={result.package.links.npm}
-                                                className="text-blue-500"
-                                                target="_blank"
-                                                rel="noreferrer"
-                                            >
-                                                <ArrowUpRightIcon className="inline-block h-4 w-4" />
-                                            </a>
-                                        </h2>
-                                        <p className="pt-2 text-sm text-gray-800">
-                                            {result.package.description
-                                                ? result.package.description
-                                                : "No description available"}
-                                        </p>
-                                        <div className="tags my-6 flex w-full flex-wrap gap-3">
-                                            {result.package.keywords?.map(keyword => {
-                                                return (
-                                                    <div
-                                                        key={keyword}
-                                                        className="tag rounded-lg bg-gray-300/70 p-2 text-sm"
-                                                    >
-                                                        {keyword}
-                                                    </div>
-                                                );
-                                            })}
-                                        </div>
-                                        <div className="info flex items-center gap-2 text-xs text-gray-500">
-                                            <p className="version">{result.package.version}</p>{" "}
-                                            {"•"}
-                                            <p className="date">
-                                                Last Updated :-{" "}
-                                                {handleLastUpdated(
-                                                    new Date(result.package.date),
-                                                    results.time,
-                                                )}
-                                            </p>
-                                        </div>
-                                        <div className="buttons mt-4 flex gap-3">
-                                            <Button
-                                                onClick={addDependency(result.package.name)}
-                                                disabled={dependencies.includes(
-                                                    result.package.name,
-                                                )}
-                                            >
-                                                Add
-                                            </Button>
-                                            <Button
-                                                onClick={addDevDependency(result.package.name)}
-                                                disabled={devDependencies.includes(
-                                                    result.package.name,
-                                                )}
-                                            >
-                                                Add as dev
-                                            </Button>
-                                        </div>
-                                    </div>
-                                );
-                            })}
-                        </div>
+                        <Results
+                            addDependency={addDependency}
+                            addDevDependency={addDevDependency}
+                            dependencies={dependencies}
+                            devDependencies={devDependencies}
+                            results={results}
+                        />
                     </>
                 )}
             </div>
